feat: randomvariable
