+++ src/continuous/elliptical/cauchy/mod.rs
@@ -1,6 +1,10 @@
 pub mod multivariate;
 pub mod univariate;
+pub mod multivariate_params;
+pub mod univariate_params;
 
+pub use multivariate_params::*;
+pub use univariate_params::*;
 pub use multivariate::*;
 pub use univariate::*;
