+++ src/continuous/elliptical/params/exact.rs
@@ -1,5 +1,4 @@
-use crate::TransformVec;
-use crate::{DistributionError, EllipticalError, EllipticalParams};
+use crate::{DistributionError, EllipticalError, EllipticalParams, RandomVariable};
 use opensrdk_linear_algebra::{matrix::ge::sy_he::po::trf::POTRF, *};
 
 #[derive(Clone, Debug)]
@@ -37,6 +36,22 @@ impl ExactEllipticalParams {
     }
 }
 
+impl RandomVariable for ExactEllipticalParams {
+    type RestoreInfo = usize;
+
+    fn transform_vec(self) -> (Vec<f64>, Self::RestoreInfo) {
+        let n = self.mu.len();
+        ([self.mu, self.lsigma.0.vec()].concat(), n)
+    }
+
+    fn restore(v: Vec<f64>, info: Self::RestoreInfo) -> Self {
+        let n = info;
+        let mu = v[0..n].to_vec();
+        let lsigma = Matrix::from(n, v[n..n + n * n].to_vec()).unwrap();
+        Self::new(mu, lsigma).unwrap()
+    }
+}
+
 impl EllipticalParams for ExactEllipticalParams {
     fn mu(&self) -> &Vec<f64> {
         self.mu()
@@ -57,21 +72,5 @@ impl EllipticalParams for ExactEllipticalParams {
             .col_mat()
             .gemm(&self.lsigma.0, &z.col_mat(), 1.0, 1.0)?
             .vec())
-    }
-}
-
-impl TransformVec for ExactEllipticalParams {
-    type T = usize;
-
-    fn transform_vec(self) -> (Vec<f64>, Self::T) {
-        let n = self.mu.len();
-        ([self.mu, self.lsigma.0.vec()].concat(), n)
-    }
-
-    fn restore(v: Vec<f64>, info: Self::T) -> Self {
-        let n = info;
-        let mu = v[0..n].to_vec();
-        let lsigma = Matrix::from(n, v[n..n + n * n].to_vec()).unwrap();
-        Self::new(mu, lsigma).unwrap()
     }
 }
