+++ src/continuous/gamma/chi_squared/mod.rs
@@ -1,5 +1,7 @@
 pub mod multivariate;
 pub mod univariate;
+pub mod univariate_params;
 
 pub use multivariate::*;
 pub use univariate::*;
+pub use univariate_params::*;
