+++ src/continuous/gamma/chi_squared/multivariate/mod.rs
@@ -1,7 +1,13 @@
 pub mod inverse_wishart;
+pub mod inverse_wishart_params;
 pub mod normal_inverse_wishart;
+pub mod normal_inverse_wishart_params;
 pub mod wishart;
+pub mod wishart_params;
 
 pub use inverse_wishart::*;
+pub use inverse_wishart_params::*;
 pub use normal_inverse_wishart::*;
+pub use normal_inverse_wishart_params::*;
 pub use wishart::*;
+pub use wishart_params::*;
