+++ src/discrete/integer/bernoulli/mod.rs
@@ -1,3 +1,7 @@
+pub mod params;
+
+pub use params::*;
+
 use crate::*;
 use rand::Rng;
 use std::ops::{BitAnd, Mul};
@@ -30,27 +34,6 @@ impl Distribution for Bernoulli {
 }
 
 impl DiscreteDistribution for Bernoulli {}
-
-#[derive(Clone, Debug)]
-pub struct BernoulliParams {
-    p: f64,
-}
-
-impl BernoulliParams {
-    pub fn new(p: f64) -> Result<Self, DistributionError> {
-        if p < 0.0 || 1.0 < p {
-            return Err(DistributionError::InvalidParameters(
-                BernoulliError::PMustBeProbability.into(),
-            ));
-        }
-
-        Ok(Self { p })
-    }
-
-    pub fn p(&self) -> f64 {
-        self.p
-    }
-}
 
 impl<Rhs, TRhs> Mul<Rhs> for Bernoulli
 where
