+++ src/distribution/event.rs
@@ -1,0 +1,5 @@
+use std::fmt::Debug;
+
+pub trait Event: Clone + Debug + Send + Sync {}
+
+impl<T> Event for T where T: Clone + Debug + Send + Sync {}
