+++ src/nonparametric/elliptical_process/sparse/mod.rs
@@ -147,6 +147,22 @@ where
     }
 }
 
+impl<K, T> RandomVariable for SparseEllipticalProcessParams<K, T>
+where
+    K: PositiveDefiniteKernel<T>,
+    T: RandomVariable,
+{
+    type RestoreInfo = ();
+
+    fn transform_vec(self) -> (Vec<f64>, Self::RestoreInfo) {
+        todo!()
+    }
+
+    fn restore(v: Vec<f64>, info: Self::RestoreInfo) -> Self {
+        todo!()
+    }
+}
+
 impl<K, T> EllipticalParams for SparseEllipticalProcessParams<K, T>
 where
     K: PositiveDefiniteKernel<T>,
