+++ src/nonparametric/point_process/dirichlet_process/pitman_yor_process/gibbs.rs
@@ -72,6 +72,18 @@ pub enum PitmanYorGibbsSample {
     New,
 }
 
+impl RandomVariable for PitmanYorGibbsSample {
+    type RestoreInfo = bool;
+
+    fn transform_vec(self) -> (Vec<f64>, Self::RestoreInfo) {
+        todo!()
+    }
+
+    fn restore(v: Vec<f64>, info: Self::RestoreInfo) -> Self {
+        todo!()
+    }
+}
+
 #[derive(Clone, Debug)]
 pub struct PitmanYorGibbsParams<'a, G0, TH>
 where
@@ -95,6 +107,22 @@ where
         n: usize,
     ) -> Self {
         Self { base, s_inv, n }
+    }
+}
+
+impl<'a, G0, TH> RandomVariable for PitmanYorGibbsParams<'a, G0, TH>
+where
+    G0: Distribution<Value = TH, Condition = ()>,
+    TH: RandomVariable,
+{
+    type RestoreInfo = ();
+
+    fn transform_vec(self) -> (Vec<f64>, Self::RestoreInfo) {
+        todo!()
+    }
+
+    fn restore(v: Vec<f64>, info: Self::RestoreInfo) -> Self {
+        todo!()
     }
 }
