+++ src/nonparametric/point_process/dirichlet_process/stick_breaking_process.rs
@@ -54,6 +54,18 @@ impl StickBreakingProcessParams {
     }
 }
 
+impl RandomVariable for StickBreakingProcessParams {
+    type RestoreInfo = ();
+
+    fn transform_vec(self) -> (Vec<f64>, Self::RestoreInfo) {
+        todo!()
+    }
+
+    fn restore(v: Vec<f64>, info: Self::RestoreInfo) -> Self {
+        todo!()
+    }
+}
+
 impl<Rhs, TRhs> Mul<Rhs> for StickBreakingProcess
 where
     Rhs: Distribution<Value = TRhs, Condition = StickBreakingProcessParams>,
